+++ CHANGELOG.md
@@ -127,8 +127,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update iPXE building script
 - Send Info, Recv, Send, and Out messages to stdout; and others to stderr
 
-- default is now protected and can't be deleted
 - first container imported container is added to the default profile
+- grub in combination can now be set as boot method with `warewulf.grubboot: true` in 
+  `warewulf.conf`. For unknown nodes `grub.efi` and `shim.efi` will be extracted from
+  the host running warewulf. If node has container it will get these binaries from the
+  container image.
+
 ## [4.4.0] 2023-01-18
 
 ### Added